Deleted animation implementation
