--- Assets/Animations/SlidingDoor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 066678cb8fb3fff46af18f02857601d0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
