--- Assets/Animations/SlidingDoor/Door_Up.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 78e7a6bf2dc074d44bc05a5599f53ef0
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
