--- Assets/Animations/SlidingDoor/SingleSlidingDoorVertical.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4bcbd4726e88b5248bf4c444c4aa0c56
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
