--- Assets/Scripts/SlidingDoor2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 07c3746eaf66454449f1ce82358a97ec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
